+++ extensions/minimax/index.ts
@@ -7,11 +7,17 @@ import {
   minimaxMediaUnderstandingProvider,
   minimaxPortalMediaUnderstandingProvider,
 } from "./media-understanding-provider.js";
-import { buildMinimaxMusicGenerationProvider } from "./music-generation-provider.js";
+import {
+  buildMinimaxMusicGenerationProvider,
+  buildMinimaxPortalMusicGenerationProvider,
+} from "./music-generation-provider.js";
 import { registerMinimaxProviders } from "./provider-registration.js";
 import { buildMinimaxSpeechProvider } from "./speech-provider.js";
 import { createMiniMaxWebSearchProvider } from "./src/minimax-web-search-provider.js";
-import { buildMinimaxVideoGenerationProvider } from "./video-generation-provider.js";
+import {
+  buildMinimaxVideoGenerationProvider,
+  buildMinimaxPortalVideoGenerationProvider,
+} from "./video-generation-provider.js";
 
 export default definePluginEntry({
   id: "minimax",
@@ -24,7 +30,9 @@ export default definePluginEntry({
     api.registerImageGenerationProvider(buildMinimaxImageGenerationProvider());
     api.registerImageGenerationProvider(buildMinimaxPortalImageGenerationProvider());
     api.registerMusicGenerationProvider(buildMinimaxMusicGenerationProvider());
+    api.registerMusicGenerationProvider(buildMinimaxPortalMusicGenerationProvider());
     api.registerVideoGenerationProvider(buildMinimaxVideoGenerationProvider());
+    api.registerVideoGenerationProvider(buildMinimaxPortalVideoGenerationProvider());
     api.registerSpeechProvider(buildMinimaxSpeechProvider());
     api.registerWebSearchProvider(createMiniMaxWebSearchProvider());
   },